feat: Moved PostgresMessageStoreIT to Jdbc Starter
